Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultaneous Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Issue: RELENG-86
Change-Id: I9bcca58a7b1f0d69a025bf0dede452e3bd6903e7
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>