Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultaneous Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Issue: RELENG-86
Change-Id: I5197b27c4e9de85f9c9bd4847a894375444d3eeb
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>