Update patch set 1
Patch Set 1:

In Fluorine, DLUX is being removed and the managed projects committee will not be supporting any DLUX or UI features or functionality.  As a result, all projects in the Fluorine simultaneous release must remove references to DLUX as a requirement.

In order to get NEMO with UI, please checkout the stable/oxygen branch and rebuild.

Patch-set: 1